+++ test/auth.js
@@ -19,6 +19,9 @@ const should = chai.should()
 
 const host = process.env.HOST
 
+const adminPrivateKey = createFakePrivateKey(0)
+const adminStakeAddress = createRewardAddress(adminPrivateKey)
+
 const stakePrivateKey1 = createFakePrivateKey(10)
 const stakeAddress1 = createRewardAddress(stakePrivateKey1)
 
@@ -46,7 +49,8 @@ const createRegisterUserSignature = (name, email, address, privateKey) => {
     host,
     action: 'Sign up',
     name,
-    email
+    email,
+    timestamp: Date.now()
   }
 
   return createCOSESign1Signature(payload, address, privateKey)
@@ -63,7 +67,8 @@ const createRegisterUserSignature = (name, email, address, privateKey) => {
 const createLoginUserSignature = (address, privateKey) => {
   const payload = {
     host,
-    action: 'Login'
+    action: 'Login',
+    timestamp: Date.now()
   }
   return createCOSESign1Signature(payload, address, privateKey)
 }
@@ -134,6 +139,67 @@ describe('*********** AUTH ***********', () => {
           res.should.has.cookie('jwt')
           accessToken = res.body.accessToken
           refreshTokenCookie = res.get('Set-Cookie')[0]
+          done()
+        })
+    })
+    it('it should NOT verify the payload if it is expired', (done) => {
+      const payload = {
+        host: process.env.HOST,
+        action: 'Login',
+        timestamp:
+          Date.now() - process.env.PAYLOAD_VALIDITY_IN_SECONDS * 1000 - 1
+      }
+
+      const coseSign1Signature = createCOSESign1Signature(
+        payload,
+        adminStakeAddress,
+        adminPrivateKey
+      )
+      const coseKey = createCOSEKey(adminPrivateKey)
+      const loginDetailsExpiredPayload = {
+        key: Buffer.from(coseKey.to_bytes()).toString('hex'),
+        signature: Buffer.from(coseSign1Signature.to_bytes()).toString('hex')
+      }
+
+      chai
+        .request(server)
+        .post('/login')
+        .send(loginDetailsExpiredPayload)
+        .end((err, res) => {
+          res.should.have.status(422)
+          res.body.should.be.a('object')
+          res.body.should.have.property('errors').that.has.property('msg')
+          res.body.errors.should.have.property('msg').eql('EXPIRED_PAYLOAD')
+          done()
+        })
+    })
+
+    it('it should NOT verify the payload if timestamp is missing', (done) => {
+      const payload = {
+        host: process.env.HOST,
+        action: 'Login'
+      }
+
+      const coseSign1Signature = createCOSESign1Signature(
+        payload,
+        adminStakeAddress,
+        adminPrivateKey
+      )
+      const coseKey = createCOSEKey(adminPrivateKey)
+      const loginDetailsExpiredPayload = {
+        key: Buffer.from(coseKey.to_bytes()).toString('hex'),
+        signature: Buffer.from(coseSign1Signature.to_bytes()).toString('hex')
+      }
+
+      chai
+        .request(server)
+        .post('/login')
+        .send(loginDetailsExpiredPayload)
+        .end((err, res) => {
+          res.should.have.status(422)
+          res.body.should.be.a('object')
+          res.body.should.have.property('errors').that.has.property('msg')
+          res.body.errors.should.have.property('msg').eql('INVALID_PAYLOAD')
           done()
         })
     })
@@ -305,7 +371,7 @@ describe('*********** AUTH ***********', () => {
       const newAddress = createRewardAddress(newPrivateKey)
       const newCoseKey = createCOSEKey(newPrivateKey)
       const newCoseSign1 = createCOSESign1Signature(
-        { host, action: 'Reset' },
+        { host, action: 'Reset', timestamp: Date.now() },
         newAddress,
         newPrivateKey
       )