+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.0.0 ()
 
 *   Wallet address uniqueness
+*   Payload expiration checking
 
 ## v0.2.0 (Dec 21, 2022)